refactor(/HomeHeader):중복되어 있던 HomeHeader컴포넌트 단일화
