+++ src/components/Header/HomeHeader/HomeHeader.module.css
@@ -13,3 +13,20 @@
   color: #fcfcfd;
 }
 
+.gnb__button {
+  height: 40px;
+
+  border: none;
+  border-radius: 8px;
+  background: transparent;
+
+  font-size: 24px;
+  cursor: pointer;
+}
+
+.gnb__button img {
+  display: block;
+  width: 40px;
+  height: 40px;
+  object-fit: contain;
+}
