+++ src/components/Header/HomeHeader/index.jsx
@@ -1,11 +1,17 @@
-import Button from '../../UI/Button';
 import styles from './HomeHeader.module.css';
 
-function HomeHeader() {
+function HomeHeader({ onRestaurantAddButtonClick }) {
   return (
     <header className={styles.gnb}>
       <h1 className={`${styles.gnb__title} text-title`}>점심 뭐 먹지</h1>
-      <Button />
+      <button
+        type="button"
+        onClick={onRestaurantAddButtonClick}
+        className={styles.gnb__button}
+        aria-label="음식점 추가"
+      >
+        <img src="/src/images/add-button.png" alt="음식점 추가" />
+      </button>
     </header>
   );
 }
